# lib/api-zod/src/generated/types/createBidBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BestBod Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBidBodyOfferType } from "./createBidBodyOfferType";

export interface CreateBidBody {
  supplierName: string;
  supplierStore: string;
  supplierEmail: string;
  price: number;
  offerType: CreateBidBodyOfferType;
  modelName: string;
  description: string;
  warrantyMonths: number;
  deliveryDays: number;
  imageUrl?: string | null;
  isSimilarModel: boolean;
}