# lib/api-zod/src/generated/types/createRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BestBod Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRequestBodyAllowedOfferTypesItem } from "./createRequestBodyAllowedOfferTypesItem";
import type { CreateRequestBodySpecifications } from "./createRequestBodySpecifications";

export interface CreateRequestBody {
  title: string;
  brand: string;
  description: string;
  categoryId: number;
  specifications: CreateRequestBodySpecifications;
  allowedOfferTypes: CreateRequestBodyAllowedOfferTypesItem[];
  allowSimilarModels: boolean;
  consumerName: string;
  consumerEmail: string;
}