# lib/api-zod/src/generated/types/interestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BestBod Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface InterestResult {
  success: boolean;
  message: string;
  contactEmail: string;
}